# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  baby-backend_internal_models.Capsule:
    properties:
      content:
        type: string
      content_type:
        description: '"text" | "photo" | "video"'
        type: string
      created_at:
        type: string
      created_date:
        type: string
      id:
        type: integer
      is_opened:
        type: boolean
      title:
        type: string
      unlock_date:
        type: string
      user_id:
        type: integer
    type: object
  baby-backend_internal_models.CreativeWork:
    properties:
      baby_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      images:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        description: '"drawing" | "handcraft" | "lego" | "sandart" | "photo" | "video"
          | "other"'
        type: string
    type: object
...
  baby-backend_internal_models.EventAlbum:
    properties:
      cover_image:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      photo_count:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  baby-backend_internal_models.FamilyMember:
    properties:
      avatar:
        type: string
      id:
        type: integer
      is_owner:
        type: boolean
      joined_at:
        type: string
      name:
        type: string
      permission:
        description: '"edit" | "view"'
        type: string
      role:
        type: string
      space_id:
        type: integer
      user_id:
        type: integer
    type: object
  baby-backend_internal_models.FamilySpace:
    properties:
      created_at:
        type: string
      id:
        type: integer
      invite_code:
        type: string
      members:
        items:
          $ref: '#/definitions/baby-backend_internal_models.FamilyMember'
        type: array
      name:
        type: string
      owner_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  baby-backend_internal_models.InviteRecord:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      permission:
        type: string
      role:
        type: string
      space_id:
        type: integer
      status:
        description: '"active" | "used" | "expired"'
        type: string
      used_at:
        type: string
      used_by:
        type: integer
    type: object
  baby-backend_internal_models.Milestone:
    properties:
      baby_id:
        type: integer
      created_at:
        type: string
      date:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      type:
        description: '"milestone" | "event"'
        type: string
    type: object
...
  baby-backend_internal_services.CreateAlbumInput:
    properties:
      cover_image:
        type: string
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  baby-backend_internal_services.CreateCapsuleInput:
    properties:
      content:
        type: string
      content_type:
        enum:
        - text
        - photo
        - video
        type: string
      created_date:
        type: string
      title:
        type: string
      unlock_date:
        type: string
    required:
    - content
    - content_type
    - created_date
    - title
    - unlock_date
    type: object
  baby-backend_internal_services.CreateCreativeWorkInput:
    properties:
      baby_id:
        type: integer
      date:
        type: string
      description:
        type: string
      image_url:
        type: string
      images:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        enum:
        - drawing
        - handcraft
        - lego
        - sandart
        - photo
        - video
        - other
        type: string
    required:
    - baby_id
    - date
    - title
    - type
    type: object
...
  baby-backend_internal_services.CreateInviteInput:
    properties:
      code:
        type: string
      permission:
        enum:
        - edit
        - view
        type: string
      role:
        type: string
    required:
    - code
    - permission
    type: object
  baby-backend_internal_services.CreateMilestoneInput:
    properties:
      baby_id:
        type: integer
      date:
        type: string
      description:
        type: string
      title:
        type: string
      type:
        enum:
        - milestone
        - event
        type: string
    required:
    - baby_id
    - date
    - title
    - type
    type: object
  baby-backend_internal_services.CreateSpaceInput:
    properties:
      invite_code:
        type: string
      name:
        type: string
      owner_name:
        type: string
    required:
    - invite_code
    - name
    - owner_name
    type: object
...
  baby-backend_internal_utils.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: 宝宝时光 · 家庭馆 — 宝宝成长记录与家庭共享后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com
  title: Baby Backend API
  version: "1.0"
paths:
  /albums:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.EventAlbum'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 相册列表
      tags:
      - Album
    post:
      consumes:
      - application/json
      parameters:
      - description: 相册信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateAlbumInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.EventAlbum'
              type: object
      security:
      - BearerAuth: []
      summary: 创建相册
      tags:
      - Album
  /albums/{id}/photos:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 相册 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 动态 ID 列表
        in: body
        name: input
        required: true
        schema:
          properties:
            entry_ids:
              items:
                type: integer
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 从相册移除照片
      tags:
      - Album
    post:
      consumes:
      - application/json
      parameters:
      - description: 相册 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 动态 ID 列表
        in: body
        name: input
        required: true
        schema:
          properties:
            entry_ids:
              items:
                type: integer
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 添加照片到相册
      tags:
      - Album
...
  /babies/{id}/milestones:
    get:
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.Milestone'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 里程碑列表
      tags:
      - Milestone
    post:
      consumes:
      - application/json
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 里程碑
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateMilestoneInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.Milestone'
              type: object
      security:
      - BearerAuth: []
      summary: 添加里程碑
      tags:
      - Milestone
  /babies/{id}/works:
    get:
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.CreativeWork'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 创意作品列表
      tags:
      - CreativeWork
    post:
      consumes:
      - application/json
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 作品信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateCreativeWorkInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.CreativeWork'
              type: object
      security:
      - BearerAuth: []
      summary: 添加创意作品
      tags:
      - CreativeWork
  /babies/{id}/works/{wid}:
    delete:
      parameters:
      - description: 宝宝 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 作品 ID
        in: path
        name: wid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除创意作品
      tags:
      - CreativeWork
  /capsules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.Capsule'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 时间胶囊列表
      tags:
      - Capsule
    post:
      consumes:
      - application/json
      parameters:
      - description: 胶囊内容
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateCapsuleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.Capsule'
              type: object
      security:
      - BearerAuth: []
      summary: 创建时间胶囊
      tags:
      - Capsule
  /capsules/{id}/open:
    post:
      parameters:
      - description: 胶囊 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 开启胶囊
      tags:
      - Capsule
  /family:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/baby-backend_internal_models.FamilySpace'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 家庭空间列表
      tags:
      - Family
    post:
      consumes:
      - application/json
      parameters:
      - description: 空间信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateSpaceInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.FamilySpace'
              type: object
      security:
      - BearerAuth: []
      summary: 创建家庭空间
      tags:
      - Family
  /family/{id}/invites:
    post:
      consumes:
      - application/json
      parameters:
      - description: 空间 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 邀请码信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/baby-backend_internal_services.CreateInviteInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/baby-backend_internal_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/baby-backend_internal_models.InviteRecord'
              type: object
      security:
      - BearerAuth: []
      summary: 创建邀请码
      tags:
      - Family
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"